Minor fixes to Converters (#427)

+++ src/main/java/com/databricks/jdbc/api/impl/fake/EmptyResultSet.java
@@ -72,7 +72,7 @@ public class EmptyResultSet implements ResultSet, IDatabricksResultSet {
   @Override
   public double getDouble(int columnIndex) throws SQLException {
     checkIfClosed();
-    return (double) 0;
+    return 0;
   }
 
   @Override
